Extract the flow matrix from a spatial interaction model object in data frame format

Extract the flow matrix from a spatial interaction model object in data frame format

Usage

flows_df(sim, ...)

Arguments

sim

a spatial interaction model object

...

additional parameters (not used currently)

Value

a data frame of flows between origin locations and destination locations with additional content if available (see Details).

Details

This function extracts the flow matrix in a long format. Each row contains the flow between an origin location and a destination location. The resulting data frame has at least three columns:

origin_idx: identifies the origin location by its index from 1 to the number of origin locations

destination_idx: identifies the destination location by its index from 1 to the number of destination locations

flow: the flow between the corresponding location

In addition, if location information is available, it will be included in the data frame as follows:

location names are included using columns origin_name or destination_name

positions are included using 2 or 3 columns (per location type, origin or destination) depending on the number of dimensions used for the location. The names of the columns are by default origin_x, origin_y and origin_z ( and equivalent names for destination location) unless coordinate names are specified in the location positions. In this latter case, the names are prefixed by origin_ or destination_. For instance, if the destination location position coordinates are named "longitude" and "latitude", the resulting columns will be destination_longitude and destination_latitude.

See also

Examples

positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
model <- static_blvim(distances, production, 1.5, 1, attractiveness)
flows_df(model)
#>     origin_idx destination_idx        flow origin_name destination_name
#> 1            1               1 0.509903048           1                1
#> 2            2               1 0.116316253           2                1
#> 3            3               1 0.155213313           3                1
#> 4            4               1 0.139760475           4                1
#> 5            5               1 0.303139143           5                1
#> 6            6               1 0.438381925           6                1
#> 7            7               1 0.117495177           7                1
#> 8            8               1 0.248888006           8                1
#> 9            9               1 0.144653843           9                1
#> 10          10               1 0.279329703          10                1
#> 11           1               2 0.017821152           1                2
#> 12           2               2 0.416008550           2                2
#> 13           3               2 0.027606752           3                2
#> 14           4               2 0.024940161           4                2
#> 15           5               2 0.034495115           5                2
#> 16           6               2 0.020384969           6                2
#> 17           7               2 0.126210789           7                2
#> 18           8               2 0.044427315           8                2
#> 19           9               2 0.017432396           9                2
#> 20          10               2 0.039462781          10                2
#> 21           1               3 0.027194652           1                3
#> 22           2               3 0.031569993           2                3
#> 23           3               3 0.363783577           3                3
#> 24           4               3 0.008167986           4                3
#> 25           5               3 0.073062539           5                3
#> 26           6               3 0.024894159           6                3
#> 27           7               3 0.015022591           7                3
#> 28           8               3 0.021013556           8                3
#> 29           9               3 0.222604558           9                3
#> 30          10               3 0.024377849          10                3
#> 31           1               4 0.017208498           1                4
#> 32           2               4 0.020042989           2                4
#> 33           3               4 0.005740095           3                4
#> 34           4               4 0.517653329           4                4
#> 35           5               4 0.012360021           5                4
#> 36           6               4 0.020892431           6                4
#> 37           7               4 0.052382422           7                4
#> 38           8               4 0.035921957           8                4
#> 39           9               4 0.004924456           9                4
#> 40          10               4 0.027291929          10                4
#> 41           1               5 0.080743112           1                5
#> 42           2               5 0.059968847           2                5
#> 43           3               5 0.111071770           3                5
#> 44           4               5 0.026737701           4                5
#> 45           5               5 0.239295292           5                5
#> 46           6               5 0.077648689           6                5
#> 47           7               5 0.040395502           7                5
#> 48           8               5 0.067541027           8                5
#> 49           9               5 0.087780204           9                5
#> 50          10               5 0.078938969          10                5
#> 51           1               6 0.150085384           1                6
#> 52           2               6 0.045551266           2                6
#> 53           3               6 0.048643933           3                6
#> 54           4               6 0.058091973           4                6
#> 55           5               6 0.099805947           5                6
#> 56           6               6 0.186170927           6                6
#> 57           7               6 0.048964708           7                6
#> 58           8               6 0.104285972           8                6
#> 59           9               6 0.043839768           9                6
#> 60          10               6 0.115048379          10                6
#> 61           1               7 0.018857585           1                7
#> 62           2               7 0.132210865           2                7
#> 63           3               7 0.013761201           3                7
#> 64           4               7 0.068279919           4                7
#> 65           5               7 0.024340841           5                7
#> 66           6               7 0.022954268           6                7
#> 67           7               7 0.397128988           7                7
#> 68           8               7 0.060132463           8                7
#> 69           9               7 0.009434118           9                7
#> 70          10               7 0.045913082          10                7
#> 71           1               8 0.070289633           1                8
#> 72           2               8 0.081892101           2                8
#> 73           3               8 0.033871336           3                8
#> 74           4               8 0.082392691           4                8
#> 75           5               8 0.071612940           5                8
#> 76           6               8 0.086025547           6                8
#> 77           7               8 0.105810867           7                8
#> 78           8               8 0.225688958           8                8
#> 79           9               8 0.026394274           9                8
#> 80          10               8 0.164051476          10                8
#> 81           1               9 0.022518174           1                9
#> 82           2               9 0.017711890           2                9
#> 83           3               9 0.197780157           3                9
#> 84           4               9 0.006225910           4                9
#> 85           5               9 0.051302239           5                9
#> 86           6               9 0.019933597           6                9
#> 87           7               9 0.009150366           7                9
#> 88           8               9 0.014548747           8                9
#> 89           9               9 0.409443917           9                9
#> 90          10               9 0.017057587          10                9
#> 91           1              10 0.085378762           1               10
#> 92           2              10 0.078727245           2               10
#> 93           3              10 0.042527865           3               10
#> 94           4              10 0.067749854           4               10
#> 95           5              10 0.090585924           5               10
#> 96           6              10 0.102713488           6               10
#> 97           7              10 0.087438590           7               10
#> 98           8              10 0.177551998           8               10
#> 99           9              10 0.033492467           9               10
#> 100         10              10 0.208528245          10               10