Compute a collection of Boltzmann–Lotka–Volterra model solutions

This function computes a collection of flows between origin locations and destination locations using blvim() on a grid of parameters. The flows use the same costs, same production constraints and same attractivenesses. Each flow is computed using one of all the pairwise combinations between the alpha values given by alphas and the beta values given by betas. The function returns an object of class sim_list which contains the resulting flows.

Usage

grid_blvim(
  costs,
  X,
  alphas,
  betas,
  Z,
  epsilon = 0.01,
  iter_max = 50000,
  conv_check = 100,
  precision = 1e-06,
  quadratic = FALSE,
  progress = FALSE
)

Arguments

costs

a cost matrix

X

a vector of production constraints

alphas

a vector of return to scale parameters

betas

a vector of cost inverse scale parameters

Z

a vector of initial destination attractivenesses

epsilon

the update intensity

iter_max

the maximal number of steps of the BLV dynamic

conv_check

number of iterations between to convergence test

precision

convergence threshold

quadratic

selects the update rule, see details.

progress

if TRUE, a progress bar is shown during the calculation (defaults to FALSE)

Value

an object of class sim_list

Examples

positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
all_flows <- grid_blvim(distances, production, c(1.25, 1.5), c(1, 2, 3), attractiveness)
all_flows
#> Collection of 6 spatial interaction models with 10 origin locations and 10
#> destination locations computed on the following grid:
#> • alpha: 1.25 and 1.5
#> • beta: 1, 2, and 3
length(all_flows)
#> [1] 6
all_flows[[2]]
#> Spatial interaction model with 10 origin locations and 10 destination locations
#> • Model: Wilson's production constrained
#> • Parameters: return to scale (alpha) = 1.5 and inverse cost scale (beta) = 1
#> ℹ The BLV model converged after 1800 iterations.