Compute an equilibrium solution of the Boltzmann–Lotka–Volterra model

This function computes flows between origin locations and destination locations at an equilibrium solution of A. Wilson's Boltzmann–Lotka–Volterra (BLV) interaction model. The BLV dynamic model is initialised with the production constraints at the origin locations and the initial values of the the attractiveness of destination locations. Iterations update the attractivenesses according the received flows.

Usage

blvim(
  costs,
  X,
  alpha,
  beta,
  Z,
  epsilon = 0.01,
  iter_max = 50000,
  conv_check = 100,
  precision = 1e-06,
  quadratic = FALSE
)

Arguments

costs

a cost matrix

X

a vector of production constraints

alpha

the return to scale parameter

beta

the cost inverse scale parameter

Z

a vector of initial destination attractivenesses

epsilon

the update intensity

iter_max

the maximal number of steps of the BLV dynamic

conv_check

number of iterations between to convergence test

precision

convergence threshold

quadratic

selects the update rule, see details.

Value

an object of class sim for spatial interaction model that contains the matrix of flows between the origin and the destination locations as well as the final attractivenesses computed by the model.

Details

In a step of the BLV model, flows are computed according to the production constrained entropy maximising model proposed by A. Wilson and implemented in static_blvim(). Then the flows received at a destination are computed as follows

$$\forall i,\quad D_j=\sum_{i=1}^{n}Y_{ij},$$

for destination \(j\). This enables updating the attractivenesses by making them closer to the received flows, i.e. trying to reduce \(|D_j-Z_j|\).

A. Wilson and co-authors proposed two different update strategies:

The original model proposed in Harris & Wilson (1978) updates the \(Z_j\) as follows $$Z_j^{t+1} = Z_j^{t} + \epsilon (D^{t}_j-Z^{t}_j)$$

In Wilson (2008), the update is given by $$Z_j^{t+1} = Z_j^{t} + \epsilon (D^{t}_j-Z^{t}_j)Z^{t}_j$$

In both cases, \(\epsilon\) is given by the epsilon parameter. It should be smaller than 1. The first update is used when the quadratic parameter is FALSE which is the default value. The second update is used when quadratic is TRUE.

Updates are performed until convergence or for a maximum of iter_max iterations. Convergence is checked every conv_check iterations. The algorithm is considered to have converged if $$\|Z^{t+1}-Z^t\|<\delta (\|Z^{t+1}\|+\delta),$$ where \(\delta\) is given by the precision parameter.

References

Harris, B., & Wilson, A. G. (1978). "Equilibrium Values and Dynamics of Attractiveness Terms in Production-Constrained Spatial-Interaction Models", Environment and Planning A: Economy and Space, 10(4), 371-388. doi:10.1068/a100371

Wilson, A. (2008), "Boltzmann, Lotka and Volterra and spatial structural evolution: an integrated methodology for some dynamical systems", J. R. Soc. Interface.5865–871 doi:10.1098/rsif.2007.1288

Examples

positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
flows <- blvim(distances, production, 1.5, 1, attractiveness)