Extract the cost matrix used to compute this model

Extract the cost matrix used to compute this model

Usage

costs(sim, ...)

Arguments

sim

a spatial interaction model with a cost matrix

...

additional parameters

Value

the cost matrix

Examples

positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
model <- static_blvim(distances, production, 1.5, 1, attractiveness)
costs(model) ## should be equal to distances above
#>            1        2         3         4         5         6         7
#> 1  0.0000000 2.109253 0.8050224 1.6680040 1.6519403 0.8412281 1.2659470
#> 2  2.1092534 0.000000 2.0790640 3.1459963 3.7380497 2.3015558 3.1772170
#> 3  0.8050224 2.079064 0.0000000 1.1146933 1.8423051 1.6387939 1.1428440
#> 4  1.6680040 3.145996 1.1146933 0.0000000 1.5309973 2.4582213 0.7874179
#> 5  1.6519403 3.738050 1.8423051 1.5309973 0.0000000 1.9862497 0.7986296
#> 6  0.8412281 2.301556 1.6387939 2.4582213 1.9862497 0.0000000 1.9070484
#> 7  1.2659470 3.177217 1.1428440 0.7874179 0.7986296 1.9070484 0.0000000
#> 8  0.1167550 2.112979 0.6919900 1.5603812 1.6295345 0.9578845 1.1925495
#> 9  1.1709448 2.105324 0.3697511 1.0486644 2.0823721 1.9974833 1.3211304
#> 10 2.7153987 3.163338 1.9132502 1.5564235 3.0861366 3.5520390 2.3165011
#>            8         9       10
#> 1  0.1167550 1.1709448 2.715399
#> 2  2.1129786 2.1053239 3.163338
#> 3  0.6919900 0.3697511 1.913250
#> 4  1.5603812 1.0486644 1.556423
#> 5  1.6295345 2.0823721 3.086137
#> 6  0.9578845 1.9974833 3.552039
#> 7  1.1925495 1.3211304 2.316501
#> 8  0.0000000 1.0595549 2.600269
#> 9  1.0595549 0.0000000 1.560583
#> 10 2.6002688 1.5605830 0.000000