Extract the inverse cost matrix used to compute this model

Extract the inverse cost matrix used to compute this model

Usage

costs(sim, ...)

Arguments

sim

a spatial interaction model with a cost matrix

...

additional parameters

Value

the cost matrix

Examples

positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
model <- static_blvim(distances, production, 1.5, 1, attractiveness)
costs(model) ## should be equal to distances above
#>            1         2         3         4         5        6         7
#> 1  0.0000000 1.2553625 0.8165432 0.4700871 1.5744235 2.355346 0.6564916
#> 2  1.2553625 0.0000000 1.6799990 0.9576644 2.1944476 3.527538 1.8383243
#> 3  0.8165432 1.6799990 0.0000000 0.7474588 0.7651751 1.903863 0.5252246
#> 4  0.4700871 0.9576644 0.7474588 0.0000000 1.3942441 2.570795 0.8984791
#> 5  1.5744235 2.1944476 0.7651751 1.3942441 0.0000000 1.887840 1.2063279
#> 6  2.3553460 3.5275375 1.9038634 2.5707952 1.8878404 0.000000 1.7071027
#> 7  0.6564916 1.8383243 0.5252246 0.8984791 1.2063279 1.707103 0.0000000
#> 8  1.4829488 0.3782033 2.0112239 1.2691649 2.5618213 3.814237 2.1092534
#> 9  0.7664458 1.9349055 1.2838965 1.2343460 2.0089234 2.089718 0.8050224
#> 10 1.8729279 3.0369908 2.1920519 2.3326359 2.7808284 1.973950 1.6680040
#>            8         9       10
#> 1  1.4829488 0.7664458 1.872928
#> 2  0.3782033 1.9349055 3.036991
#> 3  2.0112239 1.2838965 2.192052
#> 4  1.2691649 1.2343460 2.332636
#> 5  2.5618213 2.0089234 2.780828
#> 6  3.8142371 2.0897183 1.973950
#> 7  2.1092534 0.8050224 1.668004
#> 8  0.0000000 2.0790640 3.145996
#> 9  2.0790640 0.0000000 1.114693
#> 10 3.1459963 1.1146933 0.000000