Extract the inverse cost matrix used to compute this model

Extract the inverse cost matrix used to compute this model

Usage

costs(sim, ...)

Arguments

sim

a spatial interaction model with a cost matrix

...

additional parameters

Value

the cost matrix

Examples

positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
model <- static_blvim(distances, production, 1.5, 1, attractiveness)
costs(model) ## should be equal to distances above
#>           1         2         3         4         5         6         7
#> 1  0.000000 2.4708821 2.8083360 2.1583248 3.5846592 2.8604053 2.0775051
#> 2  2.470882 0.0000000 2.0674025 0.4218693 1.1541588 0.3925020 1.7599639
#> 3  2.808336 2.0674025 0.0000000 1.7494463 2.2915947 2.2425239 0.7322067
#> 4  2.158325 0.4218693 1.7494463 0.0000000 1.4275348 0.7844264 1.3621813
#> 5  3.584659 1.1541588 2.2915947 1.4275348 0.0000000 0.8195934 2.3527949
#> 6  2.860405 0.3925020 2.2425239 0.7844264 0.8195934 0.0000000 2.0425884
#> 7  2.077505 1.7599639 0.7322067 1.3621813 2.3527949 2.0425884 0.0000000
#> 8  1.655420 0.8888917 1.7491220 0.5125530 1.9375066 1.2760734 1.1767949
#> 9  1.744203 1.1591376 1.3223380 0.7373530 2.0223175 1.5061250 0.7218848
#> 10 3.098302 1.1000631 3.1549032 1.4898873 1.5021912 0.9491491 2.8510380
#>            8         9        10
#> 1  1.6554199 1.7442030 3.0983021
#> 2  0.8888917 1.1591376 1.1000631
#> 3  1.7491220 1.3223380 3.1549032
#> 4  0.5125530 0.7373530 1.4898873
#> 5  1.9375066 2.0223175 1.5021912
#> 6  1.2760734 1.5061250 0.9491491
#> 7  1.1767949 0.7218848 2.8510380
#> 8  0.0000000 0.4549828 1.8312262
#> 9  0.4549828 0.0000000 2.2037700
#> 10 1.8312262 2.2037700 0.0000000