Extract the inverse cost matrix used to compute this model

Extract the inverse cost matrix used to compute this model

Usage

costs(sim, ...)

Arguments

sim

a spatial interaction model with a cost matrix

...

additional parameters

Value

the cost matrix

Examples

positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
model <- static_blvim(distances, production, 1.5, 1, attractiveness)
costs(model) ## should be equal to distances above
#>            1         2         3         4         5        6        7        8
#> 1  0.0000000 1.3132573 1.0748930 0.9021471 0.9129873 2.703477 3.136289 1.889933
#> 2  1.3132573 0.0000000 0.5903285 0.4681089 1.1177218 3.485952 2.143499 3.168746
#> 3  1.0748930 0.5903285 0.0000000 0.3022482 0.5480141 2.905439 2.070278 2.956052
#> 4  0.9021471 0.4681089 0.3022482 0.0000000 0.6963580 3.053366 2.312575 2.789471
#> 5  0.9129873 1.1177218 0.5480141 0.6963580 0.0000000 2.368469 2.337849 2.656324
#> 6  2.7034767 3.4859516 2.9054394 3.0533659 2.3684688 0.000000 3.913146 3.048106
#> 7  3.1362894 2.1434991 2.0702777 2.3125750 2.3378488 3.913146 0.000000 4.985052
#> 8  1.8899330 3.1687460 2.9560516 2.7894715 2.6563237 3.048106 4.985052 0.000000
#> 9  1.2242554 0.7664537 0.2287887 0.5304388 0.5092932 2.790153 1.914236 3.078269
#> 10 2.0852352 1.3520282 1.9084198 1.6715538 2.3636742 4.679251 3.165795 3.506311
#>            9       10
#> 1  1.2242554 2.085235
#> 2  0.7664537 1.352028
#> 3  0.2287887 1.908420
#> 4  0.5304388 1.671554
#> 5  0.5092932 2.363674
#> 6  2.7901530 4.679251
#> 7  1.9142355 3.165795
#> 8  3.0782691 3.506311
#> 9  0.0000000 2.109667
#> 10 2.1096671 0.000000