Extract the cost matrix used to compute this model

Extract the cost matrix used to compute this model

Usage

costs(sim, ...)

Arguments

sim

a spatial interaction model with a cost matrix

...

additional parameters

Value

the cost matrix

Examples

positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
model <- static_blvim(distances, production, 1.5, 1, attractiveness)
costs(model) ## should be equal to distances above
#>            1         2         3        4         5        6         7
#> 1  0.0000000 1.1925495 1.3211304 2.316501 0.8055599 2.551823 1.9200999
#> 2  1.1925495 0.0000000 1.0595549 2.600269 0.8845568 1.420838 1.1787539
#> 3  1.3211304 1.0595549 0.0000000 1.560583 0.5164343 2.283101 0.7114704
#> 4  2.3165011 2.6002688 1.5605830 0.000000 1.7985287 3.821769 1.9565093
#> 5  0.8055599 0.8845568 0.5164343 1.798529 0.0000000 2.280105 1.1453935
#> 6  2.5518225 1.4208385 2.2831006 3.821769 2.2801054 0.000000 1.9518406
#> 7  1.9200999 1.1787539 0.7114704 1.956509 1.1453935 1.951841 0.0000000
#> 8  1.4499246 1.5326705 0.5112804 1.067611 0.7675796 2.794379 1.0940947
#> 9  3.1004639 2.9376507 1.9326500 1.171765 2.3808575 3.827040 1.8936283
#> 10 0.5780662 0.9850948 0.7481489 1.868811 0.2438174 2.404650 1.3889313
#>            8        9        10
#> 1  1.4499246 3.100464 0.5780662
#> 2  1.5326705 2.937651 0.9850948
#> 3  0.5112804 1.932650 0.7481489
#> 4  1.0676106 1.171765 1.8688109
#> 5  0.7675796 2.380858 0.2438174
#> 6  2.7943787 3.827040 2.4046501
#> 7  1.0940947 1.893628 1.3889313
#> 8  0.0000000 1.650867 0.9035321
#> 9  1.6508666 0.000000 2.5490050
#> 10 0.9035321 2.549005 0.0000000