Extract the flow matrix from a spatial interaction model object

Extract the flow matrix from a spatial interaction model object

Usage

flows(sim, ...)

Arguments

sim

a spatial interaction model object

...

additional parameters

Value

a matrix of flows between origin locations and destination locations

Examples

positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
model <- static_blvim(distances, production, 1.5, 1, attractiveness)
flows(model)
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.63598360 0.01157462 0.02748223 0.02018431 0.09340779 0.04578211
#>  [2,] 0.06084459 0.41789902 0.03208652 0.05387674 0.01643532 0.02958127
#>  [3,] 0.10021795 0.02225874 0.28990082 0.15124845 0.06065952 0.17002159
#>  [4,] 0.08054100 0.04089677 0.16550097 0.31721891 0.04032471 0.11056351
#>  [5,] 0.34308986 0.01148385 0.06109846 0.03711872 0.29199858 0.10348326
#>  [6,] 0.15841046 0.01947105 0.16132387 0.09587309 0.09748400 0.27507048
#>  [7,] 0.32510662 0.03207448 0.03913893 0.03640043 0.06606256 0.05727092
#>  [8,] 0.08582471 0.20342413 0.02789695 0.04140640 0.01993002 0.02935564
#>  [9,] 0.16054494 0.07174476 0.05969460 0.07262909 0.04545979 0.06893587
#> [10,] 0.04768248 0.01927474 0.13690742 0.13367380 0.03149217 0.08265466
#>             [,7]       [,8]       [,9]       [,10]
#>  [1,] 0.09591006 0.01756911 0.04419450 0.007911685
#>  [2,] 0.04974088 0.21890477 0.10381909 0.016811796
#>  [3,] 0.04210569 0.02082512 0.05992392 0.082838198
#>  [4,] 0.04284972 0.03382267 0.07977840 0.088503325
#>  [5,] 0.07158443 0.01498545 0.04596464 0.019192768
#>  [6,] 0.05846024 0.02079299 0.06566061 0.047453205
#>  [7,] 0.26947444 0.04933947 0.11361210 0.011520044
#>  [8,] 0.07110424 0.38834579 0.12087321 0.011838917
#>  [9,] 0.12175627 0.08988677 0.28879144 0.020556480
#> [10,] 0.02048247 0.01460626 0.03410438 0.479121617