Extract the flow matrix from a spatial interaction model object

Extract the flow matrix from a spatial interaction model object

Usage

flows(sim, ...)

Arguments

sim

a spatial interaction model object

...

additional parameters

Value

a matrix of flows between origin locations and destination locations

Examples

positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
model <- static_blvim(distances, production, 1.5, 1, attractiveness)
flows(model)
#>            [,1]       [,2]       [,3]       [,4]       [,5]        [,6]
#>  [1,] 0.4838010 0.04874440 0.07559647 0.10689700 0.03542895 0.016225869
#>  [2,] 0.2268206 0.28140670 0.05244694 0.10800058 0.03135436 0.008266932
#>  [3,] 0.2695149 0.04018312 0.21560459 0.10210354 0.10031056 0.032123336
#>  [4,] 0.3517097 0.07636375 0.09422766 0.19897367 0.04934954 0.015216453
#>  [5,] 0.1869009 0.03554622 0.14842892 0.07912566 0.31902879 0.048300453
#>  [6,] 0.1253404 0.01372365 0.06960212 0.03572545 0.07072634 0.467153751
#>  [7,] 0.3017043 0.03271697 0.12163123 0.08374188 0.06155250 0.037304560
#>  [8,] 0.2050625 0.21883973 0.04274786 0.08978142 0.02464864 0.007044913
#>  [9,] 0.3140653 0.03451591 0.06618338 0.06954540 0.03205337 0.029565482
#> [10,] 0.1772708 0.01956798 0.04555104 0.03957705 0.02528110 0.056652455
#>             [,7]       [,8]       [,9]      [,10]
#>  [1,] 0.08871786 0.03882266 0.07948013 0.02628570
#>  [2,] 0.04476724 0.19278926 0.04064580 0.01350163
#>  [3,] 0.12751341 0.02885324 0.05971299 0.02408030
#>  [4,] 0.08101979 0.05592478 0.05790632 0.01930833
#>  [5,] 0.09548350 0.02461756 0.04279229 0.01977568
#>  [6,] 0.08473728 0.01030286 0.05779719 0.06489096
#>  [7,] 0.20565876 0.02495227 0.09194549 0.03879201
#>  [8,] 0.03875615 0.31943152 0.03994402 0.01374323
#>  [9,] 0.10683691 0.02988217 0.23896709 0.07838500
#> [10,] 0.07692888 0.01754711 0.13377939 0.40784425