Extract the flow matrix from a spatial interaction model object

Extract the flow matrix from a spatial interaction model object

Usage

flows(sim, ...)

Arguments

sim

a spatial interaction model object

...

additional parameters

Value

a matrix of flows between origin locations and destination locations

Examples

positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
model <- static_blvim(distances, production, 1.5, 1, attractiveness)
flows(model)
#>             [,1]        [,2]        [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.70540918 0.054304469 0.032183166 0.01145205 0.04468652 0.01024326
#>  [2,] 0.22436239 0.364305389 0.184708597 0.02501180 0.02329380 0.01508860
#>  [3,] 0.16620803 0.230885083 0.455380427 0.01836696 0.01496835 0.01027908
#>  [4,] 0.05273181 0.027875325 0.016375832 0.40601356 0.03214201 0.17327965
#>  [5,] 0.17561642 0.022157178 0.011390410 0.02743293 0.34652913 0.03937415
#>  [6,] 0.04205552 0.014994062 0.008171763 0.15450521 0.04113463 0.36202295
#>  [7,] 0.26561953 0.085147715 0.043173791 0.04337659 0.06311624 0.03452325
#>  [8,] 0.03530453 0.008283395 0.004300403 0.06176117 0.05599330 0.14456732
#>  [9,] 0.10461529 0.019447057 0.009864308 0.04648672 0.14677143 0.07152302
#> [10,] 0.08879620 0.024248963 0.012523649 0.08469125 0.08118917 0.11150241
#>             [,7]        [,8]       [,9]      [,10]
#>  [1,] 0.07816125 0.006773039 0.03032932 0.02645775
#>  [2,] 0.10351874 0.006565624 0.02329356 0.02985150
#>  [3,] 0.06561074 0.004260745 0.01476922 0.01927137
#>  [4,] 0.05877280 0.054557959 0.06205631 0.11619474
#>  [5,] 0.07298964 0.042216076 0.16722367 0.09507039
#>  [6,] 0.04170884 0.113869725 0.08513309 0.13640420
#>  [7,] 0.29965368 0.017801226 0.06738301 0.08020498
#>  [8,] 0.02730413 0.459618973 0.10195101 0.10091577
#>  [9,] 0.06839342 0.067464785 0.30414700 0.16128697
#> [10,] 0.07920896 0.064976108 0.15693085 0.29593245