Extract the flow matrix from a spatial interaction model object

Extract the flow matrix from a spatial interaction model object

Usage

flows(sim, ...)

Arguments

sim

a spatial interaction model object

...

additional parameters

Value

a matrix of flows between origin locations and destination locations

Examples

positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
model <- static_blvim(distances, production, 1.5, 1, attractiveness)
flows(model)
#>             1          2           3          4          5          6
#> 1  0.53416792 0.11995245 0.004838279 0.11901759 0.02984380 0.04763878
#> 2  0.36380059 0.20250796 0.006509263 0.16664006 0.04132048 0.05463355
#> 3  0.03734417 0.01656569 0.515370722 0.01373856 0.08115151 0.04856118
#> 4  0.37889403 0.17491689 0.005666512 0.21256631 0.03598641 0.04734444
#> 5  0.12041408 0.05497108 0.042421677 0.04560948 0.26940832 0.12423181
#> 6  0.16920466 0.06398186 0.022346472 0.05282190 0.10936070 0.23715893
#> 7  0.25071846 0.09867188 0.013245745 0.08148770 0.07919853 0.12572139
#> 8  0.07102307 0.03623500 0.089031691 0.03117015 0.13059681 0.06022196
#> 9  0.08597374 0.03451350 0.075815429 0.02840088 0.13715181 0.11970388
#> 10 0.18900095 0.05379823 0.016267698 0.04664518 0.05612421 0.11905916
#>             7          8          9         10
#> 1  0.07805677 0.01119068 0.01797871 0.03731502
#> 2  0.09316910 0.01731569 0.02188950 0.03221380
#> 3  0.03182970 0.10827637 0.12237203 0.02479007
#> 4  0.08076495 0.01563517 0.01890736 0.02931792
#> 5  0.09948659 0.08302577 0.11572238 0.04470882
#> 6  0.13902250 0.03370262 0.08891040 0.08348995
#> 7  0.21446852 0.02656807 0.04888786 0.06103185
#> 8  0.05249611 0.42377044 0.08353405 0.02192072
#> 9  0.07278345 0.06294019 0.31929725 0.06341987
#> 10 0.09624121 0.01749412 0.06717354 0.33819570