Extract the flow matrix from a spatial interaction model object

Extract the flow matrix from a spatial interaction model object

Usage

flows(sim, ...)

Arguments

sim

a spatial interaction model object

...

additional parameters

Value

a matrix of flows between origin locations and destination locations

Examples

positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
model <- static_blvim(distances, production, 1.5, 1, attractiveness)
flows(model)
#>             [,1]       [,2]       [,3]       [,4]        [,5]       [,6]
#>  [1,] 0.76239635 0.02277956 0.01625515 0.03113775 0.007478906 0.01543041
#>  [2,] 0.05635107 0.23574778 0.02982539 0.15460805 0.074336621 0.15921579
#>  [3,] 0.06455627 0.04788250 0.37847596 0.06580569 0.038265849 0.04019042
#>  [4,] 0.07406369 0.14865974 0.03941250 0.22667774 0.054379900 0.10345151
#>  [5,] 0.02833575 0.11385250 0.03650568 0.08661974 0.361066634 0.15908985
#>  [6,] 0.04196726 0.17505010 0.02752378 0.11829102 0.114203379 0.25919334
#>  [7,] 0.10606207 0.05151355 0.14396959 0.07667897 0.028474640 0.03883106
#>  [8,] 0.12710875 0.09672292 0.04092005 0.14091955 0.033893917 0.06567174
#>  [9,] 0.11947497 0.07582678 0.06440885 0.11561158 0.031985146 0.05359545
#> [10,] 0.04786384 0.12482072 0.01599120 0.08452559 0.083491964 0.14515360
#>             [,7]       [,8]       [,9]      [,10]
#>  [1,] 0.03375878 0.05148682 0.04711271 0.01216355
#>  [2,] 0.04056066 0.09691853 0.07396743 0.07846866
#>  [3,] 0.18198892 0.06582703 0.10086818 0.01613917
#>  [4,] 0.05805252 0.13577191 0.10843776 0.05109272
#>  [5,] 0.03433851 0.05201624 0.04778659 0.08038851
#>  [6,] 0.03361545 0.07234894 0.05748070 0.10032603
#>  [7,] 0.29940848 0.09229721 0.14546332 0.01730111
#>  [8,] 0.07252615 0.23527197 0.14927039 0.03769457
#>  [9,] 0.11741378 0.15333212 0.24167386 0.02667746
#> [10,] 0.02166945 0.06008231 0.04139548 0.37500585