Extract the flow matrix from a spatial interaction model object

Extract the flow matrix from a spatial interaction model object

Usage

flows(sim, ...)

Arguments

sim

a spatial interaction model object

...

additional parameters

Value

a matrix of flows between origin locations and destination locations

Examples

positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
model <- static_blvim(distances, production, 1.5, 1, attractiveness)
flows(model)
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.70053918 0.01935836 0.06180130 0.04898510 0.03551253 0.01886274
#>  [2,] 0.06114287 0.27657927 0.08219152 0.03217128 0.04465945 0.15490892
#>  [3,] 0.20017764 0.08428839 0.28363538 0.05235301 0.05479191 0.06860623
#>  [4,] 0.14375935 0.02989256 0.04743466 0.25698897 0.15883904 0.04225518
#>  [5,] 0.09648116 0.03841467 0.04595781 0.14704358 0.23790486 0.05855815
#>  [6,] 0.05047181 0.13123312 0.05667477 0.03852584 0.05767275 0.23430775
#>  [7,] 0.14749416 0.04088086 0.06633028 0.13914594 0.13631177 0.05385241
#>  [8,] 0.05885856 0.06370870 0.04459219 0.07102547 0.11415607 0.10977909
#>  [9,] 0.03443903 0.12158818 0.04022034 0.02940695 0.04619750 0.15987723
#> [10,] 0.06777157 0.09618785 0.06581398 0.05428707 0.07911010 0.14315319
#>             [,7]       [,8]        [,9]      [,10]
#>  [1,] 0.05723575 0.02122026 0.009318725 0.02716605
#>  [2,] 0.05010600 0.07254659 0.103914001 0.12178011
#>  [3,] 0.08337241 0.05207362 0.035250817 0.08545059
#>  [4,] 0.15846564 0.07514970 0.023352184 0.06386272
#>  [5,] 0.14370992 0.11181523 0.033961345 0.08615328
#>  [6,] 0.05591674 0.10590218 0.115754091 0.15354094
#>  [7,] 0.22565758 0.08007607 0.028071097 0.08217983
#>  [8,] 0.08618947 0.24288537 0.072331022 0.13647406
#>  [9,] 0.04025746 0.09637406 0.323621160 0.10801809
#> [10,] 0.07955708 0.12274738 0.072916039 0.21845575