Extract the flow matrix from a spatial interaction model object

Extract the flow matrix from a spatial interaction model object

Usage

flows(sim, ...)

Arguments

sim

a spatial interaction model object

...

additional parameters

Value

a matrix of flows between origin locations and destination locations

Examples

positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
model <- static_blvim(distances, production, 1.5, 1, attractiveness)
flows(model)
#>             [,1]       [,2]        [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.53645833 0.09120760 0.014489861 0.04889293 0.05160628 0.04210426
#>  [2,] 0.31792222 0.23374151 0.011836388 0.04132279 0.03061538 0.02539296
#>  [3,] 0.09806064 0.02298052 0.453812545 0.12970074 0.04255356 0.06797498
#>  [4,] 0.19237920 0.04664573 0.075409153 0.26385061 0.05197370 0.06807639
#>  [5,] 0.20551320 0.03497736 0.025040475 0.05260277 0.26704419 0.16079134
#>  [6,] 0.15949864 0.02759653 0.038049550 0.06554134 0.15295247 0.25402529
#>  [7,] 0.26611643 0.19515379 0.010580139 0.03658636 0.02569238 0.02120276
#>  [8,] 0.24046441 0.05770674 0.049887630 0.17371653 0.05439411 0.06460128
#>  [9,] 0.24243768 0.13344646 0.007032699 0.02457968 0.02820515 0.02029341
#> [10,] 0.13899369 0.02391383 0.038805069 0.05844078 0.15220684 0.21812349
#>             [,7]       [,8]        [,9]      [,10]
#>  [1,] 0.06746273 0.06722738 0.047141922 0.03340870
#>  [2,] 0.17244837 0.05623566 0.090449064 0.02003565
#>  [3,] 0.01815157 0.09438842 0.009254643 0.06312239
#>  [4,] 0.03649418 0.19109468 0.018805961 0.05527040
#>  [5,] 0.02593783 0.06055979 0.021841008 0.14569203
#>  [6,] 0.02036177 0.06841754 0.014948343 0.19860853
#>  [7,] 0.26451710 0.04923312 0.114189882 0.01672804
#>  [8,] 0.04464309 0.23985604 0.023516767 0.05121340
#>  [9,] 0.14887189 0.03381161 0.344856840 0.01646458
#> [10,] 0.01764302 0.05956820 0.013319668 0.27898542