Turn a collection of spatial interaction models into a data frame

This function extracts from a collection of spatial interaction models (represented by a sim_list) a data frame in a long format, with one flow per row. This can be seen a collection oriented version of fortify.sim(). The resulting data frame is used by autoplot.sim_list() to produce summary graphics.

Usage

# S3 method for class 'sim_list'
fortify(
  model,
  data,
  flows = c("full", "destination", "attractiveness"),
  with_names = FALSE,
  normalisation = c("origin", "full", "none"),
  ...
)

Arguments

model

a collection of spatial interaction models, a sim_list

data

not used

flows

"full" (default), "destination" or "attractiveness", see details.

with_names

specifies whether the extracted data frame includes location names (FALSE by default), see details.

normalisation

when flows="full", the flows can be reported without normalisation (normalisation="none") or they can be normalised, either to sum to one for each origin location (normalisation="origin", the default value) or to sum to one globally (normalisation="full").

...

additional parameters, not used currently

Details

The data frame produced by the method depends on the values of flows and to a lesser extent on the value of with_names. In all cases, the data frame has a configuration column that identify from which spatial interaction model the other values have been extracted: this is the index of the model in the original sim_list. Depending on flows we have the following representations:

if flows="full": this is the default case for which the full flow matrix of each spatial interaction model is extracted. The data frame contains 4 columns:

origin_idx: identifies the origin location by its index from 1 to the number of origin locations

destination_idx: identifies the destination location by its index from 1 to the number of destination locations

flow: the flow between the corresponding location. By default, flows are normalised by origin location (when normalisation="origin"): the total flows originating from each origin location is normalised to 1. If normalisation="full", this normalisation is global: the sum of all flows in each model is normalised to 1. If normalisation="none" flows are not normalised.

configuration: the spatial interaction model index

if flows="destination" or flows="attractiveness", the data frame contains 3 or 4 columns:

destination: identifies the destination location by its index from 1 to the number of destination locations

flow or attractiveness depending on the value of "flows": this contains either the destination_flow() or the attractiveness() of the destination location

configuration: the spatial interaction model index

name: the destination location names if with_names=TRUE (the column is not present if with_names=FALSE)

The normalisation operated when flows="full" can improve the readability of the graphical representation proposed in autoplot.sim_list() when the production constraints differ significantly from one origin location to another.

See also

Examples

positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
flows_1 <- blvim(distances, production, 1.5, 1, attractiveness)
flows_2 <- blvim(distances, production, 1.25, 2, attractiveness)
all_flows <- sim_list(list(flows_1, flows_2))
ggplot2::fortify(all_flows) ## somewhat similar to a row bind of sim_df results
#>     origin_idx destination_idx         flow configuration
#> 1            1               1 1.000000e+00             1
#> 2            2               1 1.000000e+00             1
#> 3            3               1 1.000000e+00             1
#> 4            4               1 1.000000e+00             1
#> 5            5               1 1.000000e+00             1
#> 6            6               1 1.000000e+00             1
#> 7            7               1 1.000000e+00             1
#> 8            8               1 1.000000e+00             1
#> 9            9               1 1.000000e+00             1
#> 10          10               1 1.000000e+00             1
#> 11           1               2 9.580683e-12             1
#> 12           2               2 2.025307e-10             1
#> 13           3               2 6.680559e-11             1
#> 14           4               2 4.251867e-11             1
#> 15           5               2 1.307486e-11             1
#> 16           6               2 2.305995e-11             1
#> 17           7               2 1.401242e-11             1
#> 18           8               2 1.828387e-10             1
#> 19           9               2 1.254138e-11             1
#> 20          10               2 1.549355e-10             1
#> 21           1               3 3.062607e-12             1
#> 22           2               3 2.135539e-11             1
#> 23           3               3 3.671996e-10             1
#> 24           4               3 1.893431e-10             1
#> 25           5               3 3.062648e-12             1
#> 26           6               3 4.322949e-12             1
#> 27           7               3 3.121180e-12             1
#> 28           8               3 2.009946e-11             1
#> 29           9               3 3.077457e-12             1
#> 30          10               3 1.786698e-11             1
#> 31           1               4 1.951391e-12             1
#> 32           2               4 8.660191e-12             1
#> 33           3               4 1.206431e-10             1
#> 34           4               4 3.108498e-10             1
#> 35           5               4 2.032608e-12             1
#> 36           6               4 2.370222e-12             1
#> 37           7               4 1.954737e-12             1
#> 38           8               4 7.931511e-12             1
#> 39           9               4 2.119368e-12             1
#> 40          10               4 6.816978e-12             1
#> 41           1               5 1.808494e-12             1
#> 42           2               5 2.468072e-12             1
#> 43           3               5 1.808519e-12             1
#> 44           4               5 1.883765e-12             1
#> 45           5               5 1.987548e-11             1
#> 46           6               5 3.572693e-12             1
#> 47           7               5 8.789538e-12             1
#> 48           8               5 2.767565e-12             1
#> 49           9               5 1.904867e-11             1
#> 50          10               5 3.513211e-12             1
#> 51           1               6 2.926062e-12             1
#> 52           2               6 7.042801e-12             1
#> 53           3               6 4.130213e-12             1
#> 54           4               6 3.554089e-12             1
#> 55           5               6 5.780455e-12             1
#> 56           6               6 1.019646e-11             1
#> 57           7               6 6.195088e-12             1
#> 58           8               6 7.760064e-12             1
#> 59           9               6 5.545407e-12             1
#> 60          10               6 8.863642e-12             1
#> 61           1               7 2.935841e-12             1
#> 62           2               7 4.293871e-12             1
#> 63           3               7 2.991990e-12             1
#> 64           4               7 2.940875e-12             1
#> 65           5               7 1.426860e-11             1
#> 66           6               7 6.215792e-12             1
#> 67           7               7 1.532196e-11             1
#> 68           8               7 4.771085e-12             1
#> 69           9               7 1.370732e-11             1
#> 70          10               7 5.829125e-12             1
#> 71           1               8 1.301424e-11             1
#> 72           2               8 2.483650e-10             1
#> 73           3               8 8.541065e-11             1
#> 74           4               8 5.289693e-11             1
#> 75           5               8 1.991587e-11             1
#> 76           6               8 3.451442e-11             1
#> 77           7               8 2.114966e-11             1
#> 78           8               8 4.112395e-10             1
#> 79           9               8 1.909012e-11             1
#> 80          10               8 3.483453e-10             1
#> 81           1               9 6.559216e-13             1
#> 82           2               9 8.586199e-13             1
#> 83           3               9 6.591021e-13             1
#> 84           4               9 7.123838e-13             1
#> 85           5               9 6.908748e-12             1
#> 86           6               9 1.243088e-12             1
#> 87           7               9 3.062470e-12             1
#> 88           8               9 9.621478e-13             1
#> 89           9               9 1.120993e-11             1
#> 90          10               9 1.231828e-12             1
#> 91           1              10 6.854973e-13             1
#> 92           2              10 1.108563e-11             1
#> 93           3              10 3.999131e-12             1
#> 94           4              10 2.394713e-12             1
#> 95           5              10 1.331658e-12             1
#> 96           6              10 2.076512e-12             1
#> 97           7              10 1.361058e-12             1
#> 98           8              10 1.834835e-11             1
#> 99           9              10 1.287371e-12             1
#> 100         10              10 5.932652e-11             1
#> 101          1               1 6.856004e-01             2
#> 102          2               1 2.371309e-02             2
#> 103          3               1 8.836876e-03             2
#> 104          4               1 6.687173e-03             2
#> 105          5               1 2.380157e-02             2
#> 106          6               1 3.214145e-01             2
#> 107          7               1 1.078669e-01             2
#> 108          8               1 8.912235e-03             2
#> 109          9               1 2.585703e-02             2
#> 110         10               1 1.226715e-02             2
#> 111          1               2 2.370705e-29             2
#> 112          2               2 3.664235e-28             2
#> 113          3               2 1.485723e-29             2
#> 114          4               2 4.554235e-30             2
#> 115          5               2 1.532828e-30             2
#> 116          6               2 6.438664e-29             2
#> 117          7               2 7.978618e-30             2
#> 118          8               2 1.122370e-28             2
#> 119          9               2 1.532086e-30             2
#> 120         10               2 1.109326e-28             2
#> 121          1               3 4.224579e-03             2
#> 122          2               3 7.104493e-03             2
#> 123          3               3 7.827680e-01             2
#> 124          4               3 1.574968e-01             2
#> 125          5               3 1.466661e-04             2
#> 126          6               3 3.945987e-03             2
#> 127          7               3 6.903284e-04             2
#> 128          8               3 2.365296e-03             2
#> 129          9               3 1.608765e-04             2
#> 130         10               3 2.572622e-03             2
#> 131          1               4 3.317873e-03             2
#> 132          2               4 2.260181e-03             2
#> 133          3               4 1.634572e-01             2
#> 134          4               4 8.211903e-01             2
#> 135          5               4 1.249722e-04             2
#> 136          6               4 2.294794e-03             2
#> 137          7               4 5.237994e-04             2
#> 138          8               4 7.125226e-04             2
#> 139          9               4 1.476018e-04             2
#> 140         10               4 7.244823e-04             2
#> 141          1               5 2.312805e-01             2
#> 142          2               5 1.489832e-02             2
#> 143          3               5 2.981111e-03             2
#> 144          4               5 2.447538e-03             2
#> 145          5               5 9.697823e-01             2
#> 146          6               5 4.231461e-01             2
#> 147          7               5 8.595158e-01             2
#> 148          8               5 7.040698e-03             2
#> 149          9               5 9.677015e-01             2
#> 150         10               5 1.561657e-02             2
#> 151          1               6 2.958210e-23             2
#> 152          2               6 5.927477e-24             2
#> 153          3               6 7.596866e-25             2
#> 154          4               6 4.256867e-25             2
#> 155          5               6 4.007935e-24             2
#> 156          6               6 1.684051e-22             2
#> 157          7               6 2.086287e-23             2
#> 158          8               6 2.704632e-24             2
#> 159          9               6 4.007158e-24             2
#> 160         10               6 4.856900e-24             2
#> 161          1               7 4.603786e-25             2
#> 162          2               7 3.406161e-26             2
#> 163          3               7 6.163082e-27             2
#> 164          4               7 4.505829e-27             2
#> 165          5               7 3.775265e-25             2
#> 166          6               7 9.674699e-25             2
#> 167          7               7 1.972858e-24             2
#> 168          8               7 1.580520e-26             2
#> 169          9               7 3.784972e-25             2
#> 170         10               7 3.247353e-26             2
#> 171          1               8 7.557666e-02             2
#> 172          2               8 9.520239e-01             2
#> 173          3               8 4.195675e-02             2
#> 174          4               8 1.217818e-02             2
#> 175          5               8 6.144457e-03             2
#> 176          6               8 2.491986e-01             2
#> 177          7               8 3.140321e-02             2
#> 178          8               8 9.809692e-01             2
#> 179          9               8 6.133029e-03             2
#> 180         10               8 9.688192e-01             2
#> 181          1               9 6.822835e-31             2
#> 182          2               9 4.043709e-32             2
#> 183          3               9 8.879611e-33             2
#> 184          4               9 7.849835e-33             2
#> 185          5               9 2.627812e-30             2
#> 186          6               9 1.148839e-30             2
#> 187          7               9 2.340033e-30             2
#> 188          8               9 1.908359e-32             2
#> 189          9               9 7.515780e-30             2
#> 190         10               9 4.305600e-32             2
#> 191          1              10 2.067304e-28             2
#> 192          2              10 1.869952e-27             2
#> 193          3              10 9.068842e-29             2
#> 194          4              10 2.460771e-29             2
#> 195          5              10 2.708401e-29             2
#> 196          6              10 8.893161e-28             2
#> 197          7              10 1.282222e-28             2
#> 198          8              10 1.925317e-27             2
#> 199          9              10 2.749844e-29             2
#> 200         10              10 2.770532e-26             2
ggplot2::fortify(all_flows, flows = "destination")
#>    destination         flow configuration
#> 1            1 1.000000e+01             1
#> 2            1 1.224957e+00             2
#> 3            2 7.218985e-10             1
#> 4            2 7.081419e-28             2
#> 5            3 6.325114e-10             1
#> 6            3 9.614756e-01             2
#> 7            4 4.653299e-10             1
#> 8            4 9.947538e-01             2
#> 9            5 6.553600e-11             1
#> 10           5 3.494410e+00             2
#> 11           6 6.199428e-11             1
#> 12           6 2.415395e-22             2
#> 13           7 7.327645e-11             1
#> 14           7 4.249740e-24             2
#> 15           8 1.253942e-09             1
#> 16           8 3.324403e+00             2
#> 17           9 2.750424e-11             1
#> 18           9 1.443405e-29             2
#> 19          10 1.018964e-10             1
#> 20          10 3.289474e-26             2
destination_names(all_flows) <- letters[1:10]
ggplot2::fortify(all_flows, flows = "attractiveness", with_names = TRUE)
#>    destination attractiveness configuration name
#> 1            1   9.999994e+00             1    a
#> 2            1   1.224954e+00             2    a
#> 3            2   1.234809e-06             1    b
#> 4            2   2.370694e-22             2    b
#> 5            3   1.029521e-06             1    c
#> 6            3   9.614757e-01             2    c
#> 7            4   8.380480e-07             1    d
#> 8            4   9.947538e-01             2    d
#> 9            5   3.267237e-07             1    e
#> 10           5   3.494413e+00             2    e
#> 11           6   3.070530e-07             1    f
#> 12           6   4.221377e-18             2    f
#> 13           7   3.520882e-07             1    g
#> 14           7   2.086679e-19             2    g
#> 15           8   1.731713e-06             1    h
#> 16           8   3.324404e+00             2    h
#> 17           9   1.925102e-07             1    i
#> 18           9   1.170228e-23             2    i
#> 19          10   3.404508e-07             1    j
#> 20          10   4.140149e-21             2    j