Compute flows between origin and destination locations

This function computes flows between origin locations and destination locations according to the production constrained entropy maximising model proposed by A. Wilson.

Usage

static_blvim(costs, X, alpha, beta, Z)

Arguments

costs

a cost matrix

X

a vector of production constraints

alpha

the return to scale parameter

beta

the inverse cost scale parameter

Z

a vector of destination attractivenesses

Value

an object of class sim (and sim_wpc) for spatial interaction model that contains the matrix of flows from the origin locations to the destination locations (see \((Y_{ij})_{1\leq i\leq n, 1\leq j\leq p}\) above) and the attractivenesses of the destination locations.

Details

The model computes flows using the following parameters:

costs (\(c\)) is a \(n\times p\) matrix whose \((i,j )\) entry is the cost of having a "unitary" flow from origin location \(i\) to destination location \(j\)

X (\(X\)) is a vector of size \(n\) containing non negative production constraints for the \(n\) origin locations

alpha (\(\alpha\)) is a return to scale parameter that enhance (or reduce if smaller that 1) the attractivenesses of destination locations when they are larger than 1

beta (\(\beta\)) is the inverse of a cost scale parameter, i.e., costs are multiplied by beta in the model

Z (\(Z\)) is a vector of size \(p\) containing the positive attractivenesses of the \(p\) destination locations

According to Wilson's model, the flow from origin location \(i\) to destination location \(j\), \(Y_{ij}\), is given by

$$Y_{ij}=\frac{X_iZ_j^{\alpha}\exp(-\beta c_{ij})}{\sum_{k=1}^pZ_k^{\alpha}\exp(-\beta c_{ik})}.$$

The model is production constrained because

$$\forall i,\quad X_i=\sum_{j=1}^{p}Y_{ij},$$

that is the origin location \(i\) sends a total flow of exactly \(X_i\).

References

Wilson, A. (1971), "A family of spatial interaction models, and associated developments", Environment and Planning A: Economy and Space, 3(1), 1-32 doi:10.1068/a030001

Examples

positions <- matrix(rnorm(10 * 2), ncol = 2)
distances <- as.matrix(dist(positions))
production <- rep(1, 10)
attractiveness <- c(2, rep(1, 9))
model <- static_blvim(distances, production, 1.5, 1, attractiveness)